feat(#44): support for read-only mode
We want to open access to our test dataset directly on https://evitadb.io for all to play. In order to avoid server / catalog pollution we want to allow read-only access. Until https://github.com/FgForrest/evitaDB/issues/25 is finalized we plan to add simple boolean flag that will start the server in full read-only mode that doesn't allow clients to:

- create / update / delete catalog schemas
- open read-write sessions

This will effectively mean that the clients would not be able to alter the demo data and can only play with read queries.