feat(#86): Facet count per group
Some clients wants to display facet count on the group level (when the entire group listing is collapsed). The number cannot be computed as a mere sum of facet counts, because a single entity may have multiple facets in the same group. Therefore the count number must be computed by evitaDB directly during the facet computation phase.