fix: missing referenced entities handling update
Up to now when reference targeted managed entity and its body was requested to be fetched AND WAS NOT FOUND (for example because it was not yet indexed) the EntityReference was returned instead. This caused problems in GraphQL API which marks attributes of this reference as non-null and couldn't provide any. The same problems are in Java API in facet summary which doesn't return the information about such reference as well.

So the current behaviour is this:

- when reference is not fetched, the reference is provided because in order to eliminate it we would have to fetch the entity (in order to check its existence) and throw it away, this would slow us down
- when reference body is fetched, the reference is not returned at all because we could find out that the entity body doesn't exist