feat(#7): Hierarchy statistics refinement
In research phase we implemented `hierarchyStatistics` constraint the computed the entity cardinality for all nodes in the hierarchy. This solution can be used for "mega-menu" rendering but usually the application needs to display only part of the menu so we need to come up with refined design that would allow clients to fetch only the part of the menu they actually need to render and nothing more (we don't want to spend time on calculating things that would get thrown out).

## User requirements

- we need to render entire menu in single query (along with the filtered entities that are assigned to a hierarchy)
- we need to render "mega-menu" separately - i.e. get top X levels of all product categories - hierarchical entities (without actually fetching the product themselves)
- when we look at specific category (hierarchical entity):
   - we need to fetch parents to the root (specific levels or all)
   - we need to fetch children from current node (specific levels or all)
   - we need to fetch siblings of this node or any other node from previous two requirements
   - we need to fetch immediate children