feat(#93): Initialize catalogs on evitaDB startup in parallel
Currently, all catalogs are opened and loaded into memory synchronously. The goal of this task is to load them in parallel. Since the catalogs don't share any memory structure and don't write anything to disk (and if they did, they would write to different files/folders), it's easy to initialize them in parallel and use all CPU power to speed up evitaDB startup.