feat(#7): Refactoring of hierarchyParentsOfSelf and hierarchyParentsOfReference
The hierarchyParentsOfSelf has no sense now and will be removed without compensation.
The hierarchyParentsOfReference still has sense for scenarios where we want to access parents over entity references (for example for rendering breadcrumbs, or implement various decisions based on entity placement in the hierarchy tree). It will be removed in the form it has now, but it will by replaced by a new require constraint `hierarchyContent`, that might be optionally used inside `referenceContent`.