feat(#7): switching `hierarchyWithin` root node from plain integers to filtering constraint
When we change the filtering query parameter from plain `int` requiring hierarchy node primary keys to a more generic filter constraint we will be able to issue more complex queries such as: return all products assigned to subtrees of categories with attribute shortcut set to true. This will allow us to cross-cut the hierarchy tree in a different ways.