docs: added test support that verifies examples used in the documentation
We have more and more examples and code blocks in the documentation. The more examples, the greater the risk that one of them will become obsolete - not compilable or not working. The current situation will be prone to errors and the documentation will rot. Therefore, we need a test suite that would cover our backs and

- compile the examples
- run the examples
- verify the output of the examples (where possible)

This will is added in this commit.