feat(#101): Verify entity collection in catalog schema strictly
Now the entity collection is automatically created in the catalog when the first entity of this type is inserted. This behavior is not correct if we consider that it's possible to check the schema strictly, and also allows creation of unwanted entity collections by mistake or as a simple typo. This new feature requires the addition of something like `CatalogEvolutionMode', which will control whether the entity collections of new types are created on demand, or if it would require a deliberate `CatalogSchema` change.