feat(#84): Exact ordering as requested in input query
There are cases where the user wants to sort entities not by primary key, nor by attribute but by the order or arguments specified in `entityPrimaryKeyInSet` or `attributeInSet` constraint of the query.

The ordering will have following shapes:

**Filter by order of primary keys listed in filtering part of the query:**

```
orderBy(
   entityPrimaryKeyInFilter()
)
```

This requires that there is exactly one `entityPrimaryKeyInSet` in the filter.

**Filter by order of primary keys listed directly:**

```
orderBy(
   entityPrimaryKeyExact(5, 7, 9, 12, 6, 7)
)
```

**Filter by order of attribute constants listed in filtering part of the query:**

```
orderBy(
   attributeSetInFilter('code')
)
```

This requires that there is exactly one `attributeInSet('code', 'e', 'x', 'y')` in the filter.

**Filter by order of attributes listed directly:**

```
orderBy(
   attributeSetExact('code', 'a', 'b', 'c')
)
```

When there are multiple entities with the same attribute value, they will be inserted on the "same position" ordered by primary key ascending.