feat(#11): Rework sorting by multiple attributes
We need to avoid and remove data type `Multiple`. Problem with this type is that it hides inner data types (which there may be multiple) and there is also problem how to describe this type to the end users. We need to stick to analogies to existing and well established data store - primarily SQL. On the other hand we need to allow both these scenarios:

- sort by two or more attributes in "expected way" - i.e. secondary ordering decides order of elements when primary ordering cannot decide because values are the same
- sort by two or more attributes when the sorting handles "NULLs" last scenario - i.e. there is no attribute found for primary attribute in an entity

We also want to avoid situation where we would need to compare values of the entities in real-time. We already know it's slower than current masking process for pre-sorted arrays. This was the original reason why `Multiple` constraint exists in the first place.

The proposed solution discussed with @lho is:

- add support for creating multi-attribute sort indexes in EntitySchema  / Reference schema by adding so called SortableAttributeCompounds which will aggregate sort direction for multiple existing attributes (which doesn't need necessarily to be sortable themselves)

The final composition would like this:

```
attribute(nameOfCompound, DESC)
```