feat(#43): initial Java proxying mechanism implementation
Changed behaviour in Entity attribute internal handling. Now we throw correct exceptions when trying to access attributes that are not defined in the schema. Retrieving localized attributes without providing locale (when the locale is provided in the query that fetched the entity) is now behaving similarly as getting price for sale.