feat(#148): Add token to a GraphQL/REST/gRPC/Java query to track queries for a single unit of work
We want clients to provide a new grouping information - a random token that could be used to group multiple requests with a single unit of work - such as rendering a single page. This unit of work could be tracked in monitoring and used to "debug" the frontstore and evaluate whether the client is using the API correctly.

For example, we might discover that a single page is rendering dozens of overlapping queries instead of one or two. We could also track the impact of prefetching and other techniques common in the modern JS world.

The database could also track the queries for each unit of work and create e-tags that allow frontstores to securely cache data locally, radically reducing latency and increasing throughput - static but up-to-date e-shops, yay!