feat(#227): support UUID data type
We need to support the UUID datatype. This datatype is often used to generate IDs on the client side without the need to synchronize with other clients. Currently, the UUID would have to be stored as a string, which consumes 36 bytes, as opposed to 16 bytes if it were stored as a tuple of long values. The computational complexity on the server side is also much less if we work with the values in numerical representation.

This issue also reveals how complex is to add a new data type.