feat(#38): chained entity support
When the primary system sorts entities in the list using drag'n'drop, there is a serious problem of how to implement the order. There are these possible options:

- numerical priority from 1 ... N where first entity is marked with 1, last entity with N
  &rightarrow; disadvantage: if you move the last entity to the top, you have to update all entities
- float priority, where the entire possible numeric scope is divided so that the first entity occupies the lowest possible number in the scope and the last entity occupies the highest possible number in the scope
   &rightarrow; Disadvantage: If you move entities, the new order might be calculated as follows (left entity order + right entity order) / 2 ... this looks ok, but the possible number of divisions is surprisingly low, and from time to time you have to recompute all orders from scratch and re-position entities evenly
- instead of an order/priority number for each entity, the entities are organized as a chain of entities, where the last entity refers to its predecessor (linked list), this is a very powerful representation, but quite hard to manage and relational databases are quite slow to use it in more complex join queries

Evita should allow all three ordering variants.

Since evitaDB is considered as a secondary search index, we need to handle situations where the chain is not consistent because not every change in the primary database has been propagated to the secondary index. So the mechanism should require only one chain head, but allow multiple chain tails. Such a chain should behave like a tree, where the entities are ordered by the longest unbroken chain, but then partial branches of the tree are appended from top to bottom so that no entity is left out. The moments when the tree is not consistent should eventually be made consistent, but may prevail for a while. evitaDB should also allow setting strongly consistent requirements, where the chain must be fully consistent at transaction commit time.