feat(#281): new boolean marker for attribute representative
For better orientation in evitaLab it would be nice to be able to mark one or more attributes as representative attributes of the entity. Values of those attributes will be used in evitaLab instead of primary key when displaying value of the reference in the table cell. Representative attributes will be also automatically displayed along with entity data in the default entity listing view.

If there are more of such attributes they will be rendered next to each other joined by comma. If the attribute is localized, it will be displayed in the cell only if the language is selected. The order of the representative attributes will be driven by the order of the attribute in the entity schema.