feat(#299): Allow to override host name the evitaDB uses
The host name for all APIs is by default the host name of the system evitaDB is running on. Unfortunately, when evitaDB starts in a Docker container, the host system is a virtual system of the Docker host and neither the host name (nor the ports) matches the host system the developer is working with (unless the `--net=host` parameter is used, which doesn't work on Windows and Mac OS).

That's why we need an additional configuration property for each API that allows to propagate the name of the external system to the evitaDB instance, so that it could be used for generating links in Open API schemas and evitaLab.