feat(#304): Add requested property to Hierarchy and AttributeHistogram extra results
In order to visualize the current filter in the evitaLab visualization tab, we need to propagate the information about the requested parts to the extra results:

Hierarchy:
The entity within each hierarchy that is directly selected by the hierarchyWithin constraint is marked as requested.

AttributeHistogram:
The bucket containing the lower or upper threshold of the attributeBetween (inclusive) or within those thresholds will be marked as requested.