fix(#333): Having/excluding constraint should not be applied to root node
Currently, when using the `hierarchyWithin` with specified root node, the inner `having`/`excluding` constraints are applied even on the root node, but the root node is already specified by the node filter, thus multiple filter predicates are applied to the root node. This seems to not be ideal because if we requested certain root by its own filter, we don't expect to be excluded by subsequent `having` or `excluding` constraint. Thus we think the `having`/`excluding` constraints should by applied only to children nodes of the specified root node.

This would be especially useful when we want, for example, to construct subtree of certain root category and inside we want only categories with certain tag. But logically, the root node doesn't have that tag, because it groups together children that may or may not have that tag, that's why we filter them, anyway.

Currently, if we create the following query:
```
hierarchyWithin(
  'categories',
  attributeEquals('code', 'accessories')
  having(referenceHaving('tags', attributeEquals('code', 'material')))
)
```
we would want a subtree starting with the `accessories` category only with children that has tag `material` but because the `accessories` category groups different subcategories, the `accessories` category doesn't have that tag, and thus it would be also excluded, and thus entire subtree would be excluded. This is not ideal and expected.

Also, if we wanted the original behaviour, we could simply use the same filter from `having`/`excluding` inside the node filter and with explicit behaviour.