Refactor: Sorted collections and improved comparison logic for stabilizing test results
Adjusted and streamlined the sorting and comparison logic across multiple methods in different classes to ensure consistent ordering of keys/values. This is necessary as in certain cases the non-deterministic order of unordered data structures like HashMap was causing issues during function testing. By using LinkedHashMap and LinkedHashSet instead, the insertion order of keys/values is maintained which helps in stabilizing test results. Furthermore, comparison logic in compareTo for ReferenceKey and handling of Duplicate key exception improved for better error reporting. Sorting of prices added to PricesSerializer for serialized form consistency.

Also we got rid of TreeMaps which are not as performant as simple hash maps.