fix(#346): referencedTypeIndex fails to remove attribute value
ReferencedTypeIndex stores the information about referenced entity ids with the attributes of the entity reference. It allows us to find which ReducedEntityIndexes should be looked up for the filtered entity ids. The problem is that the ReferencedTypeIndex as the only index allows storing the same referenced id multiple times. It happens in this situation:

`Entity A` -> references `Entity D` using reference attribute `order` = `1`
`Entity B` -> references `Entity D` using reference attribute `order` = `1`

The ReferencedTypeIndex then index attribute `order` with value `1` and `Entity D` primary key twice. Then if the `Entity A` reference is dropped we need to still keep the indexed attribute `order` = `1` in the index because there is still valid reference for `Entity B`. But this was not implemented correctly and the index record was removed.

For this situations we need to keep `CardinalityIndex` that would allow us to store number of occurrences of the attribute value and remove the index record only if the last occurrence is removed. We also need to properly handle array values which need to be extended or shrinked with each added / removed reference value.