feat(#316): globally unique but language specific index
Currently we support `uniqueGlobally` attributes, which behave as a unique index over that attribute among all entities, regardless of the locale of the attribute, if the attribute is also `localized`. In some projects we want to use this index for URL attributes, but we only store relative URLs in the database. If the locale is bound (can be derived) to the missing part of the url - the domain/hostname - the URL will be the same for multiple locale variants and thus the globally unique constraint will be broken in this situation.

We need to optionally support two `uniqueGlobally' behaviors:

- **IN_ALL_LOCALES**: current behavior
- **LOCALE_SPECIFIC**: new behavior where the database will only check for attribute uniqueness among attributes in the same locale.

The second behavior disables the functionality of the implicit locale, which can no longer be derived from the unique attribute in the filter constraint for attributes with LOCALE_SPECIFIC uniqueness - since multiple locales can match the same content of the attribute value.