refactor: generics for TargetIndexes across multiple classes
Changed the declared type of TargetIndexes to use wildcard generics in multiple classes, in order to allow for broader use. This will permit the passing in of any object that extends the base Index class, instead of limiting it to only EntityIndex. This change was necessary to allow for more flexibility and generality in the codebase, as it no longer ties this part of the code strictly to the EntityIndex class, and allows for the passing in of different types of indexes.

Also added the inspection tool 'LombokGetterMayBeUsed' in Evita_DB_Defaults.xml for potential usage.