refactor: certificate directory assignment in ClientCertificateManager
Reorganized the assignment of certificateClientFolderPath and other file paths in ClientCertificateManager. Previously, set values of various file paths were overwritten under certain conditions. The conditional statement is improved to avoid this, and variables are set directly inside the branches to ensure they get the correct values. This modification aims to minimize possibility of path misassignment and improve code readability.