feat(#218): sort by primary key in descending order
The implicit ordering is ASC by primary key, but we have discovered that we lack the possibility to order by PK in descending fashion. Although this ordering is not widely used, we should support it. Because we have orderign constraints `entityPrimaryKeyInFilter` and `entityPrimaryKeyExact` we should provide new constraint to be used like that:

```
query(
   orderBy(
      entityPrimaryKey(DESC)
   )
)
```

And this would produce output sorted by PK in descending order.