fix(#162): add validation for managed entity references
Reworked original implementation which broke a lot of tests and I'm afraid it would break current clients as well. I think we need to postpone validation and allow multiple schemas to be defined in bulk to avoid unnecessary complexity on the client side while maintaining model consistency.

Current use case:

Product -> References Category
category -> references products (related products specified at category level)

It's a circular dependency, but it's a real use case that we need to support. The only situation to avoid the exception would be

1. create category
2. create product that references category
3. change category to reference related products

This requires quite complicated setup on the client side, which now uses the `ClassAnalyzer` facility to set up the schema.

So a new suggestion would be to validate the schema not in the mutation, but at the end of the updateSchema method. In `io.evitadb.core.EvitaSession#updateCatalogSchema` method the validation will be done after all mutations have been processed. This would allow to set up circular dependent updates in one go and still pass the validation.