fix(#387): Evita Client fails to query custom entity contract without specifying collection in Query
Remote Java client fails to retrieve product entity by this query:

```java
final Product product = session.queryOne(
                query(
                        filterBy(
                                entityPrimaryKeyInSet(productId)
                        )
                ),
                Product.class
        )
        .orElseThrow(
                () -> new IllegalStateException("Product with id " + productId + " not found.")
        );
```

The entity type is not defined in the query because it is expected to be retrieved from `Product.class` type via annotation examination. Unfortunatelly this information is not propagated to the server side and thus the query fails to be interpreted.