feat(#408): Optional histogram bucket optimization
Currently the histogram computer is performing an optimizations when there are a lots of empty buckets for user query. If so the histogram computer creates new histogram with fewer buckets.

This default behavior may not be ideal for all scenarios so we want to make this optimization optional. That is, in default state the histogram computer would return all requested buckets even if most of them would be empty. But the `attributeHistogram` and `priceHistogram` constraints would have flag to turn on the optimization.