feat(#406): System / gRPC API should provide information about evitaDB server version
We need to provide sever version in gRPC evitaLab API, so then the clients can visualize it and reason about it. We should also provide a method to detect the client version. Java client should log an error/warning if its minor version differs from the server version, as it might signal compatibility problems.

Implemented the gRPC part and the compatibility check.