refactor(#): speeding up the algorithm for finding a match at the beginning of a word
Until we finalize the https://github.com/FgForrest/evitaDB/issues/258 release with full text support, we could at least accelerate the `attributeStartsWith' constraint to use a binary search algorithm instead of an exhaustive search through all attribute values. In the future, this should be replaced by some form of RadixTree, but until then, we could at least use binary search here. This constraint is special in the sense that it's often used for auto-completion functions in web applications.