feat(#16): non-flushed data must be forgotten in case there is error in trunk incorporation
Otherwise the scenario when the entity is inserted fails to be repeated on retry, because OffsetIndex still maintains the file pointer from previous unsuccessful attempt and the entity looks like already inserted, which results in insertion failure (no entity is expected to exist at this time). We need to forget all the half-written data and retry again from the start.