refactor(#542): Limit allocations when no changes in indexes occur
There's a transactional boolean `dirty` in many places in transactional data structures, but it's not taken into account in the `io.evitadb.core.transaction.memory.TransactionalLayerProducer#createCopyWithMergedTransactionalMemory` method, which is called when a new state is created. If we had taken advantage of this flag and just returned "this" reference in cases where it's not dirty, we could save a lot of memory allocations. If nothing has changed, we can reuse the exact same instance of the object for the next generation.