fix(#537): Incorrect facet statistics difference when userFilter is used
When UserFilterFormula was not part of the computational tree, the facet impact computation didn't work well. Also the calculation of the best query plan tree was inverse and that was corrected as well. Finally the pretty printing visitor got split in normal and verbose mode. The normal mode is much more comprehensible when debugging larger trees.