fix(#557): attempt to fix memory leak
According to heapdump - current way of switching catalog reference is not safe and there are references to old catalogs trapped in "catalogAccessor" lambda in entity collection. These changes aim to clarify the process of transitioning from one catalog version to another and avoiding the reference leak.
According to heapdump - current way of switching catalog reference is not safe and there are references to old catalogs trapped in "catalogAccessor" lambda in entity collection. These changes aim to clarify the process of transitioning from one catalog version to another and avoiding the reference leak.