perf(#559): Entity fetching optimization
When the entity is fetched from the data store, it's entity body storage part is always fetched twice. This becomes a problem when a large number of entities are fetched from the data store, and when the OS file cache is cold, it can mean twice the latency. This fetching is not entirely necessary if we refactor the `enrichEntity` method into two separate ones and call the second one only conditionally.