fix(#556): Add support for comparing OffsetDateTime against DateTimeRange data type
If the client uses `attributeGreaterThan("validity", OffsetDateTime.now())`, the exception of a different type will be thrown. He/she needs to update the constraint to `attributeGreaterThan("validity", DateTimeRange.since(OffsetDateTime.now()))` or `attributeGreaterThan("validity", DateTimeRange.between(OffsetDateTime.now(), OffsetDateTime.now()))`, both of which work fine, but this approach is counterintuitive. We should handle the passing of OffsetDateTime and automatically convert it to a time (i.e. the `between` variant) internally to avoid confusion for developers.