fix(#577): Support for layering configuration
In complex build pipelines it's not easy to set up all evitaDB parameters in one shot. It would be helpful if we could layer the `evita-configuration.yml` so that we first load the full configuration with defaults and then apply shorter versions that override only part of those settings. The overriding configurations could be expected to be in the same folder as the default configuration and could be applied in alphabetical order - i.e:

1. evita-configuration.yml
2. evita-configuration-environment.yml
3. evita-configuration-project.yml