fix(#598): Selecting facet combination from different groups enables hasSense for invalid facets
Has-sense needs to be computed exactly and not derived from match-count. The problematic scenario is that when user selects facets A + B, from group 1 and the has-sense is calculated as false form facet C from group 2, then when he selects facet D from group 2, the facet C from group 2 becomes having sense because it starts from the baseline of facet D which is wrong.