feat(#570): Ensure the computation is killed quickly when timeout is exceeded
We set strict timeouts for requests made via HTTP endpoints, but these timeouts probably don't end the execution of the requests - they just interrupt the thread, which is not checked in the request evaluation. We need to cover this part with tests and introduce injection points that would check the thread interrupt state and terminate early if an interrupt is detected.