feat(#37): Cache inconsistency
Refactored work with QueryContext dividing it into two - QueryPlanningContext and QueryExecutionContext. The latter is unique per (isolated for) QueryPlan execution. This lead to simplification of the QueryPlanning and avoiding issues with multiple plans execution and result comparison for debug requirements.

The initialization of the formulas, sorters and computers is also changed and most of the initialization logic is moved to constructors. The only exception to this is calculation of costs which may rely on execution context and the fact whether the entities are prefetched.