fix(#611): SortIndex handles BigDecimal in wrong way
SortIndex internally holds Map, where key is the indexed value and value is the cardinality of the key. There is a normalizer implementation for certain types of keys (localized strings), but not for BigDecimal. But when the indexed values are BigDecimals and come in different forms - like `0` and `0.000` - they don't match in the map lookup unless normalized by `stripTrailingZeros()`. This causes the sort index to break during indexing.