fix(#609): Unique ID identifying the catalog
Catalogs are uniquely identified by their name. However, this name can change using `rename' and `replace' operations on the catalog. So if we have catalogs `a` and `b`, and later `b` replaces `a` and starts using the `a` catalog name, we cannot distinguish it from the original `a` catalog. By introducing a new internally assigned UUID for each catalog, we can still distinguish catalog `a` with UUID `1` from catalog `a` with UUID 2. This internal ID can be used by the outside world to infer the identity of the catalog and be used for synchronization purposes.