fix(#202): Make defaults explicit
We should explicitly express default values in Constraints. It would allow us to mark them as defaults in automatically generated schemas and omit them when printing query in examples in documentation - see the problem with REMOVE_EMPTY enum, that is being generated in REST/C# examples even if it's unnecessary.