feat(#187): updated reversed mutation stream logic
Going back to the reverse, we still need to lead each transaction with a transaction mutation (although according to standard logic it should come last).

The first reason is that it's much easier to understand from a user perspective, and the second is that this mutation contains information about the number of tx mutations, which is necessary for navigation and expectations about the stream properties.