feat(#479): Sticking sorting randomness
Another scenario that the client solves is the reliability of the output to the end user. Consider this situation - we have 10 products we want to promote on the first 3 positions of the listing. In order to cycle them all, we choose to sort them randomly (so every time there are different 3 promoted products out of 10 available). But this could be confusing for the user who refreshes the page or lists through additional pages. This situation can be easily solved by passing the seed from the client side to the random function, which could generate the seed per user and store it in their session cache.