feat(#479): Segmentation rules
Often the product listing page contains listings in differently sorted forms. The typical example is the "Top Sellers" block, which lists the exact same products as the main listing on the page, but sorts them differently. If we could merge these two listings into a single query that returns two results, we could benefit in two ways:

1. we wouldn't be forced to declare the same complex set of filtering constraints and requirements in two different queries.
2. the database engine could reuse the computed result of filtered products to satisfy both queries at once - just applying different sorting and paging requirements.

So we end up with a faster and simpler system for this use case, which is ultimately what we want.

The idea is to optionally define segment rules for paging requirement that could integrate different slices of the result with completely different sorting.