fix(#717): Client and server timeouts are not aligned
There are some operations that may take quite a while - for example replacing `goLive` operation (which reloads entire catalog in memory) or (`replaceCatalog`). These operations randomly times out when executed from gRPC client because the client and server timeouts are not aligned and either of those might cancel the operation any time.