chore: Set time travel to disabled by default
Although time travel works on its own, we've found that it's impractical to use in production because it takes up too much disk space. These issues need to be addressed first so that this feature doesn't cause problems on production systems:

- https://github.com/FgForrest/evitaDB/issues/760
- https://github.com/FgForrest/evitaDB/issues/421
- https://github.com/FgForrest/evitaDB/issues/761

That's why we temporarily disable this feature in the default settings.

Refs: #772